chore(core/rust): move Font and Color into their own modules
